+++ constructease/constructease/src/main/java/br/com/constructease/constructease/util/FormatadorDecimal.java
@@ -4,9 +4,8 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 public class FormatadorDecimal {
-    public static double arredondar(double valor) {
-        return BigDecimal.valueOf(valor)
-                .setScale(2, RoundingMode.HALF_UP)
-                .doubleValue();
+
+    public static BigDecimal arredondar(BigDecimal valor) {
+        return valor.setScale(2, RoundingMode.HALF_UP);
     }
 }
